pyproject.toml
[project]
name = "vectorforge-a-rag-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "groq>=0.31.0",
    "langchain-groq>=0.3.7",
    "python-dotenv>=1.1.1",
    "streamlit>=1.48.1",
]
